fix(typescript-eslint): align require-await rule
