+++ internal/plugins/typescript/rules/require_await/require_await.md
@@ -34,4 +34,4 @@ function baz() {
 ## Original Documentation
 
 - [typescript-eslint: require-await](https://typescript-eslint.io/rules/require-await)
-- [Source code](https://github.com/typescript-eslint/typescript-eslint/blob/v8.67.0/packages/eslint-plugin/src/rules/require-await.ts)
+- [Source code](https://github.com/typescript-eslint/typescript-eslint/blob/v8.68.0/packages/eslint-plugin/src/rules/require-await.ts)
