+++ internal/plugins/typescript/rules/require_await/require_await_upstream_test.go
@@ -1,0 +1,1207 @@
+// TestRequireAwaitUpstream migrates the typescript-eslint v8.68.0 require-await suite and its ESLint base-rule cases.
+// rslint-specific lock-in cases live in require_await_extras_test.go.
+package require_await
+
+import (
+	"strings"
+	"testing"
+
+	"github.com/web-infra-dev/rslint/internal/plugins/typescript/rules/fixtures"
+	"github.com/web-infra-dev/rslint/internal/rule_tester"
+)
+
+func requireAwaitLeadingSpaces(value string) int {
+	for _, line := range strings.Split(value, "\n") {
+		if strings.TrimSpace(line) == "" {
+			continue
+		}
+		return len(line) - len(strings.TrimLeft(line, " "))
+	}
+	return 0
+}
+
+func alignRequireAwaitSuggestionIndent(cases []rule_tester.InvalidTestCase) []rule_tester.InvalidTestCase {
+	for caseIndex := range cases {
+		codeIndent := requireAwaitLeadingSpaces(cases[caseIndex].Code)
+		for errorIndex := range cases[caseIndex].Errors {
+			for suggestionIndex := range cases[caseIndex].Errors[errorIndex].Suggestions {
+				suggestion := &cases[caseIndex].Errors[errorIndex].Suggestions[suggestionIndex]
+				difference := codeIndent - requireAwaitLeadingSpaces(suggestion.Output)
+				if difference <= 0 || !strings.HasPrefix(suggestion.Output, "\n") {
+					continue
+				}
+				lines := strings.Split(suggestion.Output, "\n")
+				prefix := strings.Repeat(" ", difference)
+				for lineIndex := 1; lineIndex < len(lines); lineIndex++ {
+					lines[lineIndex] = prefix + lines[lineIndex]
+				}
+				suggestion.Output = strings.Join(lines, "\n")
+			}
+		}
+	}
+	return cases
+}
+
+func TestRequireAwaitUpstream(t *testing.T) {
+	rule_tester.RunRuleTester(fixtures.GetRootDir(), "tsconfig.json", t, &RequireAwaitRule, []rule_tester.ValidTestCase{
+		{Code: `
+function numberOne(): number {
+  return 1;
+}
+    `},
+		{Code: `
+const numberOne = function (): number {
+  return 1;
+};
+    `},
+		{Code: `
+      const numberOne = (): number => 1;
+    `},
+		{Code: `
+const numberOne = (): number => {
+  return 1;
+};
+    `},
+		{Code: `
+function delay() {
+  return Promise.resolve();
+}
+    `},
+		{Code: `
+const delay = () => {
+  return Promise.resolve();
+};
+    `},
+		{Code: "const delay = () => Promise.resolve();"},
+		{Code: `
+async function numberOne(): Promise<number> {
+  return await 1;
+}
+    `},
+		{Code: `
+const numberOne = async function (): Promise<number> {
+  return await 1;
+};
+    `},
+		{Code: "const numberOne = async (): Promise<number> => await 1;"},
+		{Code: `
+const numberOne = async (): Promise<number> => {
+  return await 1;
+};
+    `},
+		{Code: `
+async function numberOne(): Promise<number> {
+  return Promise.resolve(1);
+}
+    `},
+		{Code: `
+const numberOne = async function (): Promise<number> {
+  return Promise.resolve(1);
+};
+    `},
+		{Code: "const numberOne = async (): Promise<number> => Promise.resolve(1);"},
+		{Code: `
+const numberOne = async (): Promise<number> => {
+  return Promise.resolve(1);
+};
+    `},
+		{Code: `
+async function numberOne(): Promise<number> {
+  return getAsyncNumber(1);
+}
+async function getAsyncNumber(x: number): Promise<number> {
+  return Promise.resolve(x);
+}
+    `},
+		{Code: `
+const numberOne = async function (): Promise<number> {
+  return getAsyncNumber(1);
+};
+const getAsyncNumber = async function (x: number): Promise<number> {
+  return Promise.resolve(x);
+};
+    `},
+		{Code: `
+const numberOne = async (): Promise<number> => getAsyncNumber(1);
+const getAsyncNumber = async function (x: number): Promise<number> {
+  return Promise.resolve(x);
+};
+    `},
+		{Code: `
+const numberOne = async (): Promise<number> => {
+  return getAsyncNumber(1);
+};
+const getAsyncNumber = async function (x: number): Promise<number> {
+  return Promise.resolve(x);
+};
+    `},
+		{Code: `
+async function testFunction(): Promise<void> {
+  await Promise.all(
+    [1, 2, 3].map(
+      // this should not trigger an error on the parent function
+      async value => Promise.resolve(value),
+    ),
+  );
+}
+    `},
+		{Code: `
+function* test6() {
+  yield* syncGenerator();
+}
+    `},
+		{Code: `
+function* syncGenerator() {
+  yield 1;
+}
+    `},
+		{Code: `
+async function* asyncGenerator() {
+  await Promise.resolve();
+  yield 1;
+}
+async function* test1() {
+  yield* asyncGenerator();
+}
+    `},
+		{Code: `
+async function* asyncGenerator() {
+  await Promise.resolve();
+  yield 1;
+}
+async function* test1() {
+  yield* asyncGenerator();
+  yield* 2;
+}
+    `},
+		{Code: `
+async function* test(source: AsyncIterable<any>) {
+  yield* source;
+}
+    `},
+		{Code: `
+async function* test(source: Iterable<any> & AsyncIterable<any>) {
+  yield* source;
+}
+    `},
+		{Code: `
+async function* test(source: Iterable<any> | AsyncIterable<any>) {
+  yield* source;
+}
+    `},
+		{Code: `
+type MyType = {
+  [Symbol.iterator](): Iterator<any>;
+  [Symbol.asyncIterator](): AsyncIterator<any>;
+};
+async function* test(source: MyType) {
+  yield* source;
+}
+    `},
+		{Code: `
+type MyType = {
+  [Symbol.asyncIterator]: () => AsyncIterator<any>;
+};
+async function* test(source: MyType) {
+  yield* source;
+}
+    `},
+		{Code: `
+type MyFunctionType = () => AsyncIterator<any>;
+type MyType = {
+  [Symbol.asyncIterator]: MyFunctionType;
+};
+async function* test(source: MyType) {
+  yield* source;
+}
+    `},
+		{Code: `
+async function* foo(): Promise<string> {
+  return new Promise(res => res(` + "`" + `hello` + "`" + `));
+}
+    `},
+		{Code: `
+      async function* f() {
+        let x!: Omit<
+          {
+            [Symbol.asyncIterator](): AsyncIterator<any>;
+          },
+          'z'
+        >;
+        yield* x;
+      }
+    `},
+		{Code: `
+      const fn = async () => {
+        await using foo = new Bar();
+      };
+    `},
+		{Code: `
+      async function* test1() {
+        yield Promise.resolve(1);
+      }
+    `},
+		{Code: `
+      function asyncFunction() {
+        return Promise.resolve(1);
+      }
+      async function* test1() {
+        yield asyncFunction();
+      }
+    `},
+		{Code: `
+      declare const asyncFunction: () => Promise<void>;
+      async function* test1() {
+        yield asyncFunction();
+      }
+    `},
+		{Code: `
+      async function* test1() {
+        yield new Promise(() => {});
+      }
+    `},
+	}, alignRequireAwaitSuggestionIndent([]rule_tester.InvalidTestCase{
+		{
+			Code: `
+async function numberOne(): Promise<number> {
+  return 1;
+}
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+function numberOne(): number {
+  return 1;
+}
+      `,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+const numberOne = async function (): Promise<number> {
+  return 1;
+};
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+const numberOne = function (): number {
+  return 1;
+};
+      `,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: "const numberOne = async (): Promise<number> => 1;",
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output:    "const numberOne = (): number => 1;",
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+async function values(): Promise<Array<number>> {
+  return [1];
+}
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+function values(): Array<number> {
+  return [1];
+}
+      `,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        async function foo() {
+          function nested() {
+            await doSomething();
+          }
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  function foo() {
+    function nested() {
+      await doSomething();
+    }
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+async function* foo(): void {
+  doSomething();
+}
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+function* foo(): void {
+  doSomething();
+}
+      `,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+async function* foo() {
+  yield 1;
+}
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+function* foo() {
+  yield 1;
+}
+      `,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+const foo = async function* () {
+  console.log('bar');
+};
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+const foo = function* () {
+  console.log('bar');
+};
+      `,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+async function* asyncGenerator() {
+  yield 1;
+}
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+function* asyncGenerator() {
+  yield 1;
+}
+      `,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+async function* asyncGenerator(source: Iterable<any>) {
+  yield* source;
+}
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+function* asyncGenerator(source: Iterable<any>) {
+  yield* source;
+}
+      `,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+function isAsyncIterable(value: unknown): value is AsyncIterable<any> {
+  return true;
+}
+async function* asyncGenerator(source: Iterable<any> | AsyncIterable<any>) {
+  if (!isAsyncIterable(source)) {
+    yield* source;
+  }
+}
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+function isAsyncIterable(value: unknown): value is AsyncIterable<any> {
+  return true;
+}
+function* asyncGenerator(source: Iterable<any> | AsyncIterable<any>) {
+  if (!isAsyncIterable(source)) {
+    yield* source;
+  }
+}
+      `,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+function* syncGenerator() {
+  yield 1;
+}
+async function* asyncGenerator() {
+  yield* syncGenerator();
+}
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+function* syncGenerator() {
+  yield 1;
+}
+function* asyncGenerator() {
+  yield* syncGenerator();
+}
+      `,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+async function* asyncGenerator() {
+  yield* anotherAsyncGenerator(); // Unknown function.
+}
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+function* asyncGenerator() {
+  yield* anotherAsyncGenerator(); // Unknown function.
+}
+      `,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        const fn = async () => {
+          using foo = new Bar();
+        };
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  const fn = () => {
+    using foo = new Bar();
+  };
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        // intentional TS error
+        async function* foo(): Promise<number> {
+          yield 1;
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  // intentional TS error
+  function* foo(): Promise<number> {
+    yield 1;
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        async function* foo(): AsyncGenerator {
+          yield 1;
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  function* foo(): Generator {
+    yield 1;
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        async function* foo(): AsyncGenerator<number> {
+          yield 1;
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  function* foo(): Generator<number> {
+    yield 1;
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+	}))
+}
+
+// TestRequireAwaitUpstreamEslintBase migrates the applicable ESLint v10.9.1 base-rule cases.
+func TestRequireAwaitUpstreamEslintBase(t *testing.T) {
+	rule_tester.RunRuleTester(fixtures.GetRootDir(), "tsconfig.json", t, &RequireAwaitRule, []rule_tester.ValidTestCase{
+		{Code: `
+async function foo() {
+  await doSomething();
+}
+    `},
+		{Code: `
+(async function () {
+  await doSomething();
+});
+    `},
+		{Code: `
+async () => {
+  await doSomething();
+};
+    `},
+		{Code: "async () => await doSomething();"},
+		{Code: `
+({
+  async foo() {
+    await doSomething();
+  },
+});
+    `},
+		{Code: `
+class A {
+  async foo() {
+    await doSomething();
+  }
+}
+    `},
+		{Code: `
+(class {
+  async foo() {
+    await doSomething();
+  }
+});
+    `},
+		{Code: `
+async function foo() {
+  await (async () => {
+    await doSomething();
+  });
+}
+    `},
+		{Code: "async function foo() {}"},
+		{Code: "async () => {};"},
+		{Code: `
+function foo() {
+  doSomething();
+}
+    `},
+		{Code: `
+async function foo() {
+  for await (x of xs);
+}
+    `},
+		{
+			Code: "await foo();",
+		},
+		{
+			Code: `
+for await (let num of asyncIterable) {
+  console.log(num);
+}
+			`,
+		},
+		{Code: "await using resource = getResource();"},
+		{Code: "async function run() { await using resource = getResource(); }"},
+		{
+			Code: `
+        async function* run() {
+          await new Promise(resolve => setTimeout(resolve, 100));
+          yield 'Hello';
+          console.log('World');
+        }
+      `,
+		},
+		{
+			Code: "async function* run() {}",
+		},
+		{
+			Code: "const foo = async function* () {};",
+		},
+	}, alignRequireAwaitSuggestionIndent([]rule_tester.InvalidTestCase{
+		{
+			Code: `
+        async function foo() {
+          doSomething();
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  function foo() {
+    doSomething();
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        (async function () {
+          doSomething();
+        });
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  (function () {
+    doSomething();
+  });
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        async () => {
+          doSomething();
+        };
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  () => {
+    doSomething();
+  };
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: "async () => doSomething();",
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output:    "() => doSomething();",
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        ({
+          async foo() {
+            doSomething();
+          },
+        });
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  ({
+    foo() {
+      doSomething();
+    },
+  });
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        class A {
+          async foo() {
+            doSomething();
+          }
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  class A {
+    foo() {
+      doSomething();
+    }
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        class A {
+          public async foo() {
+            doSomething();
+          }
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  class A {
+    public foo() {
+      doSomething();
+    }
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        (class {
+          async foo() {
+            doSomething();
+          }
+        });
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  (class {
+    foo() {
+      doSomething();
+    }
+  });
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        (class {
+          async ''() {
+            doSomething();
+          }
+        });
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  (class {
+    ''() {
+      doSomething();
+    }
+  });
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        async function foo() {
+          async () => {
+            await doSomething();
+          };
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  function foo() {
+    async () => {
+      await doSomething();
+    };
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        async function foo() {
+          await (async () => {
+            doSomething();
+          });
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  async function foo() {
+    await (() => {
+      doSomething();
+    });
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        const obj = {
+          async: async function foo() {
+            bar();
+          },
+        };
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  const obj = {
+    async: function foo() {
+      bar();
+    },
+  };
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        async    /* test */ function foo() {
+          doSomething();
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  /* test */ function foo() {
+    doSomething();
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        class A {
+          a = 0
+          async [b]() {
+            return 0;
+          }
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  class A {
+    a = 0
+    ;[b]() {
+      return 0;
+    }
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        foo
+        async () => {
+          return 0;
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  foo
+  ;() => {
+    return 0;
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `
+        class A {
+          foo() {}
+          async [bar]() {
+            baz;
+          }
+        }
+      `,
+			Errors: []rule_tester.InvalidTestCaseError{
+				{
+					MessageId: "missingAwait",
+					Suggestions: []rule_tester.InvalidTestCaseSuggestion{
+						{
+							MessageId: "removeAsync",
+							Output: `
+  class A {
+    foo() {}
+    [bar]() {
+      baz;
+    }
+  }
+`,
+						},
+					},
+				},
+			},
+		},
+		{
+			Code: `class A {
+  a
+  async [b]() { return 0; }
+}`,
+			Errors: []rule_tester.InvalidTestCaseError{{
+				MessageId: "missingAwait",
+				Line:      3,
+				Column:    3,
+				EndLine:   3,
+				EndColumn: 12,
+				Suggestions: []rule_tester.InvalidTestCaseSuggestion{{
+					MessageId: "removeAsync",
+					Output: `class A {
+  a
+  [b]() { return 0; }
+}`,
+				}},
+			}},
+		},
+		{
+			Code: `class A {
+  a = 0
+  async in() { return 0; }
+}`,
+			Errors: []rule_tester.InvalidTestCaseError{{
+				MessageId: "missingAwait",
+				Line:      3,
+				Column:    3,
+				EndLine:   3,
+				EndColumn: 11,
+				Suggestions: []rule_tester.InvalidTestCaseSuggestion{{
+					MessageId: "removeAsync",
+					Output: `class A {
+  a = 0
+  ;in() { return 0; }
+}`,
+				}},
+			}},
+		},
+		{
+			Code: `const obj = {
+  foo,
+  async in() { return 0; }
+};`,
+			Errors: []rule_tester.InvalidTestCaseError{{
+				MessageId: "missingAwait",
+				Line:      3,
+				Column:    3,
+				EndLine:   3,
+				EndColumn: 11,
+				Suggestions: []rule_tester.InvalidTestCaseSuggestion{{
+					MessageId: "removeAsync",
+					Output: `const obj = {
+  foo,
+  in() { return 0; }
+};`,
+				}},
+			}},
+		},
+		{
+			Code: "async function run() { using resource = getResource(); }",
+			Errors: []rule_tester.InvalidTestCaseError{{
+				MessageId: "missingAwait",
+				Line:      1,
+				Column:    1,
+				EndLine:   1,
+				EndColumn: 19,
+				Suggestions: []rule_tester.InvalidTestCaseSuggestion{{
+					MessageId: "removeAsync",
+					Output:    "function run() { using resource = getResource(); }",
+				}},
+			}},
+		},
+	}))
+}
